refactor: Delete event_json_parser_viewer - Rename btnVentanaSecundaria
