+++ src/components/newWindowBtn/btnVentanaSecundaria.tsx
